# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Contract Lifecycle Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}/roles:
    patch:
      operationId: updateUserRoles
      tags: [users]
      summary: Update user roles (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRolesBody"
      responses:
        "200":
          description: Roles updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Admin - Contract Types
  /admin/contract-types:
    post:
      operationId: createContractType
      tags: [admin]
      summary: Create a contract type (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractTypeBody"
      responses:
        "201":
          description: Contract type created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractType"
  /admin/contract-types/{id}:
    patch:
      operationId: updateContractType
      tags: [admin]
      summary: Update a contract type (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractTypeBody"
      responses:
        "200":
          description: Updated contract type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractType"

    delete:
      operationId: deleteContractType
      tags: [admin]
      summary: Delete a contract type (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

  # Admin - Workflows
  /admin/workflows:
    get:
      operationId: listWorkflows
      tags: [admin]
      summary: List all workflow definitions (admin only)
      responses:
        "200":
          description: Workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkflowDefinition"

    post:
      operationId: createWorkflow
      tags: [admin]
      summary: Create a workflow definition (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkflowBody"
      responses:
        "201":
          description: Workflow created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDefinition"
  /admin/workflows/{id}:
    patch:
      operationId: updateWorkflow
      tags: [admin]
      summary: Update a workflow definition (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkflowBody"
      responses:
        "200":
          description: Updated workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDefinition"

  # Admin - Screening Criteria
  /admin/screening-criteria:
    get:
      operationId: listScreeningCriteria
      tags: [admin]
      summary: List AI screening criteria (admin only)
      responses:
        "200":
          description: Screening criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScreeningCriterion"

    post:
      operationId: createScreeningCriterion
      tags: [admin]
      summary: Create a screening criterion (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScreeningCriterionBody"
      responses:
        "201":
          description: Criterion created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreeningCriterion"
  /admin/screening-criteria/{id}:
    patch:
      operationId: updateScreeningCriterion
      tags: [admin]
      summary: Update a screening criterion (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScreeningCriterionBody"
      responses:
        "200":
          description: Updated criterion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreeningCriterion"

  # Admin - Value Tiers
  /admin/value-tiers:
    get:
      operationId: listValueTiers
      tags: [admin]
      summary: List value tiers (admin only)
      responses:
        "200":
          description: Value tiers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValueTier"

    post:
      operationId: createValueTier
      tags: [admin]
      summary: Create a value tier (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateValueTierBody"
      responses:
        "201":
          description: Value tier created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValueTier"
  /admin/value-tiers/{id}:
    patch:
      operationId: updateValueTier
      tags: [admin]
      summary: Update a value tier (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateValueTierBody"
      responses:
        "200":
          description: Updated value tier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValueTier"

    delete:
      operationId: deleteValueTier
      tags: [admin]
      summary: Delete a value tier (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

  # Admin - Dashboard Config
  /admin/dashboard-config:
    get:
      operationId: getDashboardConfig
      tags: [admin]
      summary: Get dashboard widget configuration
      responses:
        "200":
          description: Dashboard config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardConfig"

    patch:
      operationId: updateDashboardConfig
      tags: [admin]
      summary: Update dashboard widget configuration (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DashboardConfig"
      responses:
        "200":
          description: Config updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardConfig"
components:
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    User:
      type: object
      properties:
        id:
          type: integer
        googleId:
          type: string
        email:
          type: string
        name:
          type: string
        photoUrl:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        roles:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - googleId
        - email
        - name
        - photoUrl
        - department
        - roles
        - createdAt
    UpdateRolesBody:
      type: object
      properties:
        roles:
          type: array
          items:
            type: string
      required:
        - roles
...
    ContractType:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        formSchema:
          type: object
        obligationTypes:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - formSchema
        - isActive
        - createdAt
    CreateContractTypeBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        formSchema:
          type: object
        obligationTypes:
          type: array
          items:
            type: string
        isActive:
          type: boolean
      required:
        - name
        - formSchema
    WorkflowDefinition:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        direction:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        minTierId:
          type: ["integer", "null"]
        maxTierId:
          type: ["integer", "null"]
        stages:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStage"
        isDefault:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - stages
        - isDefault
        - createdAt
    WorkflowStage:
      type: object
      properties:
        id:
          type: integer
        workflowId:
          type: integer
        name:
          type: string
        stageOrder:
          type: integer
        assignedRole:
          type: ["string", "null"]
        assignedUserId:
          type: ["integer", "null"]
        assignedUserName:
          type: ["string", "null"]
        canSendBack:
          type: boolean
        isOptional:
          type: boolean
        isSigned:
          type: boolean
      required:
        - id
        - workflowId
        - name
        - stageOrder
        - canSendBack
        - isOptional
        - isSigned
    CreateWorkflowBody:
      type: object
      properties:
        name:
          type: string
        direction:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        minTierId:
          type: ["integer", "null"]
        maxTierId:
          type: ["integer", "null"]
        isDefault:
          type: boolean
        stages:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              stageOrder:
                type: integer
              assignedRole:
                type: ["string", "null"]
              assignedUserId:
                type: ["integer", "null"]
              canSendBack:
                type: boolean
              isOptional:
                type: boolean
              isSigned:
                type: boolean
            required:
              - name
              - stageOrder
              - canSendBack
              - isOptional
              - isSigned
      required:
        - name
        - stages
...
    ScreeningCriterion:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        isEnabled:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - isEnabled
        - createdAt
    CreateScreeningCriterionBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        isEnabled:
          type: boolean
      required:
        - name
        - description
        - isEnabled
    ValueTier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        minValue:
          type: number
        maxValue:
          type: ["number", "null"]
        displayOrder:
          type: integer
      required:
        - id
        - name
        - minValue
        - displayOrder
    CreateValueTierBody:
      type: object
      properties:
        name:
          type: string
        minValue:
          type: number
        maxValue:
          type: ["number", "null"]
        displayOrder:
          type: integer
      required:
        - name
        - minValue
        - displayOrder
...
    DashboardConfig:
      type: object
      properties:
        showActiveValue:
          type: boolean
        showContractsByType:
          type: boolean
        showExpiringContracts:
          type: boolean
        showAverageCycleTime:
          type: boolean
        showBottlenecks:
          type: boolean
        showReviewerWorkload:
          type: boolean
        bottleneckThresholdDays:
          type: integer
        expiringWindowDays:
          type: integer
      required:
        - showActiveValue
        - showContractsByType
        - showExpiringContracts
        - showAverageCycleTime
        - showBottlenecks
        - showReviewerWorkload
        - bottleneckThresholdDays
        - expiringWindowDays